StickyMessage feature; simplify specific-channel features; refactor database

+++ src/features/PermaRole.ts
@@ -1,4 +1,3 @@
-import { GuildMember } from 'discord.js';
 import { log, prisma } from '../infrastructure';
 import { Feature } from '.';
 
@@ -13,38 +12,25 @@ export const PermaRole: Feature = {
       role => !currentRoles.has(role.id),
     );
     if (!addedRoles.size && !removedRoles.size) return;
-    const userSnowflake = BigInt(newMember.id);
+    const userSf = BigInt(newMember.id);
     await prisma.$transaction([
       ...addedRoles.map(({ id }) =>
-        prisma.permaRole.create({
-          data: { userSnowflake, roleSnowflake: BigInt(id) },
-        }),
+        prisma.permaRole.create({ data: { userSf, roleSf: BigInt(id) } }),
       ),
       ...removedRoles.map(({ id }) =>
-        prisma.permaRole.deleteMany({
-          where: { userSnowflake, roleSnowflake: BigInt(id) },
-        }),
+        prisma.permaRole.deleteMany({ where: { userSf, roleSf: BigInt(id) } }),
       ),
     ]);
-    log(
-      'MemberUpdate',
-      newMember.id,
-      addedRoles.map(role => role.id),
-      removedRoles.map(role => role.id),
-    );
   },
   /** Restore roles if any are stored */
-  async HandleMemberAdd(member: GuildMember) {
-    const userSnowflake = BigInt(member.id);
+  async HandleMemberAdd(member) {
+    const userSf = BigInt(member.id);
     const permaRoles = await prisma.permaRole.findMany({
-      select: { roleSnowflake: true },
-      where: { userSnowflake },
+      select: { roleSf: true },
+      where: { userSf },
     });
-    const roleSnowflakes = permaRoles.map(
-      ({ roleSnowflake }) => `${roleSnowflake}`,
-    );
+    const roleSnowflakes = permaRoles.map(({ roleSf }) => `${roleSf}`);
     await member.roles.add(roleSnowflakes);
-    log('MemberAdd', member.id, roleSnowflakes);
+    log('PermaRole restore', member.id, roleSnowflakes);
   },
-  async HandleMessageCreate() {},
 };
